requirement.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
Django==3.2
django-cors-headers==3.10.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
idna==3.3
Pillow==8.4.0
psycopg2-binary==2.9.2
pycparser==2.21
PyJWT==2.3.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
stripe==2.64.0
urllib3==1.26.7
